+++ packages/app-builder-lib/src/indexInternal.ts
@@ -7,7 +7,17 @@ export { CustomWindowsSign, WindowsSignTaskConfiguration } from "./codeSign/win/
 export { Configuration, ToolsetConfig, ToolsetCustom } from "./configuration.js"
 export { Publish } from "./core.js"
 export { getElectronVersion } from "./electron/electronVersion.js"
-export { FileMatcher, getFileMatchers, GetFileMatchersOptions } from "./fileMatcher.js"
+export {
+  collectExplicitReincludes,
+  DEFAULT_EXCLUDED_EXTENSIONS,
+  DEFAULT_EXCLUDED_NAMES,
+  FileMatcher,
+  getDefaultIgnoredPatterns,
+  getFileMatchers,
+  GetFileMatchersOptions,
+  getMainFileMatchers,
+  getReincludedDefaultExclusions,
+} from "./fileMatcher.js"
 export { hoist, HoisterDependencyKind, HoisterResult, HoisterTree } from "./node-module-collector/hoist.js"
 export {
   BunNodeModulesCollector,
@@ -28,14 +38,16 @@ export { createBlockmap } from "./targets/differentialUpdateInfoBuilder.js"
 export { validateCriticalPathString } from "./targets/linux/appimage/appImageUtil.js"
 export { copyMimeTypes } from "./targets/linux/appimage/appLauncher.js"
 export { MacTargetHelper, type PlatformType } from "./targets/mac/MacTargetHelper.js"
-export { computeArchToTargetNamesMap } from "./targets/targetFactory.js"
+export { assertSafeHelperName, getAvailableHelperSuffixes, type AvailableHelpers } from "./electron/mac/electronMacUtils.js"
+export { addTargetsForPlatform, computeArchToTargetNamesMap } from "./targets/targetFactory.js"
 export type { Defines } from "./targets/win/nsis/Defines.js"
 export { nsisEscapeString, NsisScriptGenerator } from "./targets/win/nsis/nsisScriptGenerator.js"
+export { ProgIdMaker } from "./targets/win/nsis/progId.js"
 export { checkMakensisOutput, verifyInstallerSize } from "./targets/win/nsis/nsisValidation.js"
 export { getLinuxToolsMacToolset, getLinuxToolsPath } from "./toolsets/linuxToolsMac.js"
 export { getWindowsKitsBundle } from "./toolsets/winCodeSign.js"
 export { CacheState } from "./util/cacheState.js"
-export { computeDefaultAppDirectory, doMergeConfigs, getConfig, validateConfiguration } from "./util/config/config.js"
+export { computeDefaultAppDirectory, createProjectMetadataLazy, doMergeConfigs, getConfig, validateConfiguration } from "./util/config/config.js"
 export { loadEnv, orNullIfFileNotExist } from "./util/config/load.js"
 export { validateSchema } from "./util/config/schemaValidator.js"
 export {